Arch: Suppress compile time warnings due to -Wall
There were a few compile time warnings which 'g++' was complaining about
on my system. Since I have not yet modified the sample code, I assume
these are harmless and can be ignore. Thus, slight modifications were
made to quiet my version of gcc.

+++ init.cpp
@@ -277,7 +277,7 @@ resolveHost(register char *name)
 		bcopy(fhost->h_addr, &sa.sin_addr, fhost->h_length);
 	} else {
 		fadd.s_addr = inet_addr(name);
-		if (fadd.s_addr != -1) {
+		if ((int)fadd.s_addr != -1) {
 			sa.sin_family = AF_INET;	/* grot */
 			sa.sin_port = 0;
 			sa.sin_addr.s_addr = fadd.s_addr;
